% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulljoinXL.R
\name{fulljoinXL}
\alias{fulljoinXL}
\alias{fulljoinXL()}
\title{Performs a SQL type full join on 2 Excel files}
\usage{
fulljoinXL()
}
\value{
The joined file, full_join.xlsx is placed in the working directory.
}
\description{
Merges all rows of 2 Excel files based upon a common column in each file.
Choose 2 files via dialog box. Prompted at the console to make 1 file a base file.
Prompted to choose a column from each Excel file to use as the join condition.
}
\examples{
\dontrun{
fulljoinXL();
}

}
\author{
Yvonne Glanville, \email{yvonneglanville@gmail.com}
}
\seealso{
\code{\link{innerjoinXL}} for SQL type inner join, \code{\link{leftjoinXL}} for SQL type left join,
\code{\link{rightjoinXL}} for SQL type right join, \code{\link{minusXL}} for file1 minus file2

Other X.joinXL.: \code{\link{innerjoinXL}},
  \code{\link{leftjoinXL}}, \code{\link{minusXL}},
  \code{\link{rightjoinXL}}
}

