% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerjoinXL.R
\name{innerjoinXL}
\alias{innerjoinXL}
\alias{innerjoinXL()}
\title{Performs a SQL type inner join on 2 Excel files}
\usage{
innerjoinXL()
}
\value{
The joined file, inner_join.xlsx is placed in the working directory.
}
\description{
Merges all rows from base file and join file if the join condition is met. Choose
2 files via dialog box. Prompted at the console to make 1 file a base file.
Prompted to choose a column from each Excel file to use as the join condition.
}
\author{
Yvonne Glanville, \email{yvonneglanville@gmail.com}
}
\seealso{
\code{\link{fulljoinXL}} for SQL type full join, \code{\link{leftjoinXL}} for SQL type left join,
\code{\link{rightjoinXL}} for SQL type right join, \code{\link{minusXL}} for file1 minus file2

Other X.joinXL.: \code{\link{fulljoinXL}},
  \code{\link{leftjoinXL}}, \code{\link{minusXL}},
  \code{\link{rightjoinXL}}
}

