% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftjoinXL.R
\name{leftjoinXL}
\alias{leftjoinXL}
\alias{leftjoinXL()}
\title{Performs a SQL type left join on 2 Excel files}
\usage{
leftjoinXL()
}
\value{
The joined file, left_join.xlsx is placed in the working directory.
}
\description{
Selects all rows from the base file, and all rows from the join file if the join
condition is met. Choose 2 files via dialog box. Prompted at the console to make 1 file
a base file. Prompted to choose a column from each Excel file to use as the join condition.
}
\examples{
\dontrun{
leftjoinXL();
}

}
\author{
Yvonne Glanville, \email{yvonneglanville@gmail.com}
}
\seealso{
\code{\link{innerjoinXL}} for SQL type inner join, \code{\link{fulljoinXL}} for SQL type full join,
\code{\link{rightjoinXL}} for SQL type right join, \code{\link{minusXL}} for file1 minus file2

Other X.joinXL.: \code{\link{fulljoinXL}},
  \code{\link{innerjoinXL}}, \code{\link{minusXL}},
  \code{\link{rightjoinXL}}
}

