% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minusXL.R
\name{minusXL}
\alias{minusXL}
\alias{minusXL()}
\title{Performs a SQL type minus query on 2 Excel files}
\usage{
minusXL()
}
\value{
In working directory: source_minus_target.xlsx, target_minus_source.xlsx, rowsNOTduplicated.xlsx
}
\description{
Performs 2 operations: source-minus-target and target-minus-source
If the files are identical all outputs are empty.
Choose 2 files via dialog box, and then from the console choose the common column between the files.
3 separate .xlsx files are returned to the user
   "source_minus_target.xlsx"
   "target_minus_source.xlsx"
   "rowsNOTduplicated.xlsx"
}
\examples{
\dontrun{
minusXL();
}


}
\author{
Yvonne Glanville, \email{yvonneglanville@gmail.com}
}
\seealso{
\code{\link{innerjoinXL}} for SQL type inner join, \code{\link{leftjoinXL}} for SQL type left join,
\code{\link{rightjoinXL}} for SQL type right join, \code{\link{fulljoinXL}} for SQL type full join

Other X.joinXL.: \code{\link{fulljoinXL}},
  \code{\link{innerjoinXL}}, \code{\link{leftjoinXL}},
  \code{\link{rightjoinXL}}
}

